Renamed the module name and workaround of the  BTCC exchange issue
