+++ befh/exch_bitstamp.py
@@ -1,9 +1,9 @@
-from bitcoinexchangefh.ws_api_socket import WebSocketApiClient
-from bitcoinexchangefh.market_data import L2Depth, Trade
-from bitcoinexchangefh.exchange import ExchangeGateway
-from bitcoinexchangefh.instrument import Instrument
-from bitcoinexchangefh.sql_client_template import SqlClientTemplate
-from bitcoinexchangefh.util import Logger
+from befh.ws_api_socket import WebSocketApiClient
+from befh.market_data import L2Depth, Trade
+from befh.exchange import ExchangeGateway
+from befh.instrument import Instrument
+from befh.sql_client_template import SqlClientTemplate
+from befh.util import Logger
 import time
 import threading
 import json
@@ -11,7 +11,7 @@ from functools import partial
 from datetime import datetime
 
 
-class ExchGwApiTemplate(WebSocketApiClient):
+class ExchGwApiBitstamp(WebSocketApiClient):
     """
     Exchange socket
     """
@@ -19,11 +19,7 @@ class ExchGwApiTemplate(WebSocketApiClient):
         """
         Constructor
         """
-        WebSocketApiClient.__init__(self, 'Template')
-        
-    @classmethod
-    def get_order_book_timestamp_field_name(cls):
-        return 'timestamp'
+        WebSocketApiClient.__init__(self, 'Bitstamp')
         
     @classmethod
     def get_trades_timestamp_field_name(cls):
@@ -39,11 +35,11 @@ class ExchGwApiTemplate(WebSocketApiClient):
         
     @classmethod
     def get_trade_side_field_name(cls):
-        return 'side'
+        return 'type'
         
     @classmethod
     def get_trade_id_field_name(cls):
-        return 'trdMatchID'
+        return 'id'
         
     @classmethod
     def get_trade_price_field_name(cls):
@@ -51,19 +47,25 @@ class ExchGwApiTemplate(WebSocketApiClient):
         
     @classmethod
     def get_trade_volume_field_name(cls):
-        return 'size'   
+        return 'amount'   
         
     @classmethod
     def get_link(cls):
-        return 'wss://www.bitmex.com/realtime'
-        
+        return 'ws://ws.pusherapp.com/app/de504dc5763aeef9ff52?protocol=7'
+
     @classmethod
     def get_order_book_subscription_string(cls, instmt):
-        return json.dumps({"op":"subscribe", "args": ["orderBook10:%s" % instmt.get_instmt_code()]})
-        
+        if instmt.get_instmt_code() == "":
+            return json.dumps({"event":"pusher:subscribe","data":{"channel":"order_book"}})
+        else:
+            return json.dumps({"event":"pusher:subscribe","data":{"channel":"order_book_%s" % instmt.get_instmt_code()}})
+
     @classmethod
     def get_trades_subscription_string(cls, instmt):
-        return json.dumps({"op":"subscribe", "args": ["trade:%s" % instmt.get_instmt_code()]})
+        if instmt.get_instmt_code() == "":
+            return json.dumps({"event":"pusher:subscribe","data":{"channel":"live_trades"}})        
+        else:
+            return json.dumps({"event":"pusher:subscribe","data":{"channel":"live_trades_%s" % instmt.get_instmt_code()}})        
             
     @classmethod
     def parse_l2_depth(cls, instmt, raw):
@@ -74,28 +76,25 @@ class ExchGwApiTemplate(WebSocketApiClient):
         """
         l2_depth = instmt.get_l2_depth()
         keys = list(raw.keys())
-        if cls.get_order_book_timestamp_field_name() in keys and \
-           cls.get_bids_field_name() in keys and \
+        if cls.get_bids_field_name() in keys and \
            cls.get_asks_field_name() in keys:
             
             # Date time
-            timestamp = raw[cls.get_order_book_timestamp_field_name()]
-            timestamp = timestamp.replace('T', ' ').replace('Z', '').replace('-' , '')
-            l2_depth.date_time = timestamp
+            l2_depth.date_time = datetime.utcnow().strftime("%Y%m%d %H:%M:%S.%f")
             
             # Bids
             bids = raw[cls.get_bids_field_name()]
-            bids = sorted(bids, key=lambda x: x[0], reverse=True)
             for i in range(0, len(bids)):
                 l2_depth.bids[i].price = float(bids[i][0]) if type(bids[i][0]) != float else bids[i][0]
                 l2_depth.bids[i].volume = float(bids[i][1]) if type(bids[i][1]) != float else bids[i][1]   
+            bids = sorted(bids, key=lambda x: x[0], reverse=True)
                 
             # Asks
             asks = raw[cls.get_asks_field_name()]
-            asks = sorted(asks, key=lambda x: x[0])
             for i in range(0, len(asks)):
                 l2_depth.asks[i].price = float(asks[i][0]) if type(asks[i][0]) != float else asks[i][0]
                 l2_depth.asks[i].volume = float(asks[i][1]) if type(asks[i][1]) != float else asks[i][1]            
+            asks = sorted(asks, key=lambda x: x[0])
         else:
             raise Exception('Does not contain order book keys in instmt %s-%s.\nOriginal:\n%s' % \
                 (instmt.get_exchange_name(), instmt.get_instmt_name(), \
@@ -120,12 +119,13 @@ class ExchGwApiTemplate(WebSocketApiClient):
            cls.get_trade_volume_field_name() in keys:
         
             # Date time
-            timestamp = raw[cls.get_trades_timestamp_field_name()]
-            timestamp = timestamp.replace('T', ' ').replace('Z', '').replace('-' , '')
-            trade.date_time = timestamp
+            date_time = float(raw[cls.get_trades_timestamp_field_name()])
+            trade.date_time = datetime.utcfromtimestamp(date_time).strftime("%Y%m%d %H:%M:%S.%f")            
             
             # Trade side
-            trade.trade_side = Trade.parse_side(raw[cls.get_trade_side_field_name()])
+            # Buy = 0
+            # Side = 1
+            trade.trade_side = Trade.parse_side(raw[cls.get_trade_side_field_name()] + 1)
                 
             # Trade id
             trade.trade_id = raw[cls.get_trade_id_field_name()]
@@ -143,7 +143,7 @@ class ExchGwApiTemplate(WebSocketApiClient):
         return trade        
 
 
-class ExchGwTemplate(ExchangeGateway):
+class ExchGwBitstamp(ExchangeGateway):
     """
     Exchange gateway
     """
@@ -152,7 +152,7 @@ class ExchGwTemplate(ExchangeGateway):
         Constructor
         :param db_client: Database client
         """
-        ExchangeGateway.__init__(self, ExchGwApiTemplate(), db_client)
+        ExchangeGateway.__init__(self, ExchGwApiBitstamp(), db_client)
 
     @classmethod
     def get_exchange_name(cls):
@@ -160,7 +160,7 @@ class ExchGwTemplate(ExchangeGateway):
         Get exchange name
         :return: Exchange name string
         """
-        return 'Template'
+        return 'Bitstamp'
 
     def on_open_handler(self, instmt, ws):
         """
@@ -172,7 +172,7 @@ class ExchGwTemplate(ExchangeGateway):
                   (instmt.get_instmt_code(), instmt.get_exchange_name()))
         if not instmt.get_subscribed():
             ws.send(self.api_socket.get_order_book_subscription_string(instmt))
-            ws.send(self.api_socket.get_trades_subscription_string(instmt))
+            ws.send(self.api_socket.get_trades_subscription_string(instmt))            
             instmt.set_subscribed(True)
 
     def on_close_handler(self, instmt, ws):
@@ -181,7 +181,7 @@ class ExchGwTemplate(ExchangeGateway):
         :param instmt: Instrument
         :param ws: Web socket
         """
-        Logger.info(self.__class__.__name__, "Instrument %s is subscribed in channel %s" % \
+        Logger.info(self.__class__.__name__, "Instrument %s is unsubscribed in channel %s" % \
                   (instmt.get_instmt_code(), instmt.get_exchange_name()))
         instmt.set_subscribed(False)
 
@@ -192,34 +192,22 @@ class ExchGwTemplate(ExchangeGateway):
         :param message: Message
         """
         keys = message.keys()
-        if 'info' in keys:
-            Logger.info(self.__class__.__name__, message['info'])
-        elif 'subscribe' in keys:
-            Logger.info(self.__class__.__name__, 'Subscription of %s is %s' % \
-                        (message['request']['args'], \
-                         'successful' if message['success'] else 'failed'))
-        elif 'table' in keys:
-            if message['table'] == 'trade':
-                for trade_raw in message['data']:
-                    if trade_raw["symbol"] == instmt.get_instmt_code():
-                        # Filter out the initial subscriptions
-                        trade = self.api_socket.parse_trade(instmt, trade_raw)
-                        if trade.trade_id != instmt.get_exch_trade_id():
-                            instmt.incr_trade_id()
-                            instmt.set_exch_trade_id(trade.trade_id)
-                            self.insert_trade(instmt, trade)
-            elif message['table'] == 'orderBook10':
-                for data in message['data']:
-                    if data["symbol"] == instmt.get_instmt_code():
-                        instmt.set_prev_l2_depth(instmt.get_l2_depth().copy())
-                        self.api_socket.parse_l2_depth(instmt, data)
-                        if instmt.get_l2_depth().is_diff(instmt.get_prev_l2_depth()):
-                            instmt.incr_order_book_id()
-                            self.insert_order_book(instmt)
-            else:
-                Logger.info(self.__class__.__name__, json.dumps(message,indent=2))
-        else:
-            Logger.error(self.__class__.__name__, "Unrecognised message:\n" + json.dumps(message))
+        if 'event' in keys and message['event'] in ['data', 'trade'] and 'channel' in keys and 'data' in keys:
+            channel_name = message['channel']
+            if (instmt.get_instmt_code() == "" and channel_name == "order_book") or \
+               (instmt.get_instmt_code() != "" and channel_name == "order_book_%s" % instmt.get_instmt_code()):
+                instmt.set_prev_l2_depth(instmt.get_l2_depth().copy())
+                self.api_socket.parse_l2_depth(instmt, json.loads(message['data']))
+                if instmt.get_l2_depth().is_diff(instmt.get_prev_l2_depth()):
+                    instmt.incr_order_book_id()
+                    self.insert_order_book(instmt)    
+            elif (instmt.get_instmt_code() == "" and channel_name == "live_trades") or \
+                 (instmt.get_instmt_code() != "" and channel_name == "live_trades_%s" % instmt.get_instmt_code()):
+                    trade = self.api_socket.parse_trade(instmt, json.loads(message['data']))
+                    if trade.trade_id != instmt.get_exch_trade_id():
+                        instmt.incr_trade_id()
+                        instmt.set_exch_trade_id(trade.trade_id)
+                        self.insert_trade(instmt, trade)     
 
     def start(self, instmt):
         """
@@ -227,8 +215,8 @@ class ExchGwTemplate(ExchangeGateway):
         :param instmt: Instrument
         :return List of threads
         """
-        instmt.set_l2_depth(L2Depth(10))
-        instmt.set_prev_l2_depth(L2Depth(10))
+        instmt.set_l2_depth(L2Depth(20))
+        instmt.set_prev_l2_depth(L2Depth(20))
         instmt.set_order_book_table_name(self.get_order_book_table_name(instmt.get_exchange_name(),
                                                                        instmt.get_instmt_name()))
         instmt.set_trades_table_name(self.get_trades_table_name(instmt.get_exchange_name(),
@@ -244,12 +232,12 @@ class ExchGwTemplate(ExchangeGateway):
                                         
 
 if __name__ == '__main__':
-    exchange_name = 'Template'
-    instmt_name = 'XBTUSD'
-    instmt_code = 'XBTH17'
+    exchange_name = 'Bitstamp'
+    instmt_name = 'BTCUSD'
+    instmt_code = ''
     instmt = Instrument(exchange_name, instmt_name, instmt_code)
     db_client = SqlClientTemplate()
     Logger.init_log()
-    exch = ExchGwTemplate(db_client)
+    exch = ExchGwBitstamp(db_client)
     td = exch.start(instmt)
