+++ befh/exch_btcc.py
@@ -1,18 +1,18 @@
-from bitcoinexchangefh.restful_api_socket import RESTfulApiSocket
-from bitcoinexchangefh.exchange import ExchangeGateway
-from bitcoinexchangefh.market_data import L2Depth, Trade
-from bitcoinexchangefh.util import Logger
-from bitcoinexchangefh.instrument import Instrument
-from bitcoinexchangefh.sql_client_template import SqlClientTemplate
+from befh.restful_api_socket import RESTfulApiSocket
+from befh.exchange import ExchangeGateway
+from befh.market_data import L2Depth, Trade
+from befh.util import Logger
+from befh.instrument import Instrument
+from befh.sql_client_template import SqlClientTemplate
 import time
 import threading
 from functools import partial
 from datetime import datetime
 
 
-class ExchGwApiQuoine(RESTfulApiSocket):
+class ExchGwBtccRestfulApi(RESTfulApiSocket):
     """
-    Exchange gateway RESTfulApi
+    Exchange gateway BTCC RESTfulApi
     """
     def __init__(self):
         RESTfulApiSocket.__init__(self)
@@ -22,41 +22,44 @@ class ExchGwApiQuoine(RESTfulApiSocket):
         return 1
         
     @classmethod
+    def get_order_book_timestamp_field_name(cls):
+        return ''
+        
+    @classmethod
     def get_trades_timestamp_field_name(cls):
-        return 'created_at'
-    
+        return ''
+        
     @classmethod
     def get_bids_field_name(cls):
-        return 'buy_price_levels'
+        return ''
         
     @classmethod
     def get_asks_field_name(cls):
-        return 'sell_price_levels'
+        return ''
         
     @classmethod
     def get_trade_side_field_name(cls):
-        return 'taker_side'
+        return ''
         
     @classmethod
     def get_trade_id_field_name(cls):
-        return 'id'
+        return ''
         
     @classmethod
     def get_trade_price_field_name(cls):
-        return 'price'        
+        return ''        
         
     @classmethod
     def get_trade_volume_field_name(cls):
-        return 'quantity'
+        return ''        
         
     @classmethod
     def get_order_book_link(cls, instmt):
-        return "https://api.quoine.com/products/%s/price_levels" % instmt.get_instmt_code()
+        return ""
 
     @classmethod
-    def get_trades_link(cls, instmt, page=1):
-            return "https://api.quoine.com/executions?product_id=%s&page=%d" % \
-                   (instmt.get_instmt_code(), page)
+    def get_trades_link(cls, instmt, trade_id=''):
+        return ""
                 
     @classmethod
     def parse_l2_depth(cls, instmt, raw):
@@ -67,23 +70,26 @@ class ExchGwApiQuoine(RESTfulApiSocket):
         """
         l2_depth = L2Depth()
         keys = list(raw.keys())
-        if cls.get_bids_field_name() in keys and \
+        if cls.get_order_book_timestamp_field_name() in keys and \
+           cls.get_bids_field_name() in keys and \
            cls.get_asks_field_name() in keys:
             
             # Date time
-            l2_depth.date_time = datetime.utcnow().strftime("%Y%m%d %H:%M:%S.%f")
+            date_time = float(raw[cls.get_order_book_timestamp_field_name()])
+            date_time = date_time / cls.get_timestamp_offset()
+            l2_depth.date_time = datetime.utcfromtimestamp(date_time).strftime("%Y%m%d %H:%M:%S.%f")
             
             # Bids
             bids = raw[cls.get_bids_field_name()]
             bids = sorted(bids, key=lambda x: x[0], reverse=True)
-            for i in range(0, 5):
+            for i in range(0, min(5, len(bids))):
                 l2_depth.bids[i].price = float(bids[i][0]) if type(bids[i][0]) != float else bids[i][0]
                 l2_depth.bids[i].volume = float(bids[i][1]) if type(bids[i][1]) != float else bids[i][1]   
                 
             # Asks
             asks = raw[cls.get_asks_field_name()]
             asks = sorted(asks, key=lambda x: x[0])
-            for i in range(0, 5):
+            for i in range(0, min(5, len(asks))):
                 l2_depth.asks[i].price = float(asks[i][0]) if type(asks[i][0]) != float else asks[i][0]
                 l2_depth.asks[i].volume = float(asks[i][1]) if type(asks[i][1]) != float else asks[i][1]            
         else:
@@ -106,8 +112,7 @@ class ExchGwApiQuoine(RESTfulApiSocket):
         if cls.get_trades_timestamp_field_name() in keys and \
            cls.get_trade_id_field_name() in keys and \
            cls.get_trade_price_field_name() in keys and \
-           cls.get_trade_volume_field_name() in keys and \
-           cls.get_trade_side_field_name() in keys:
+           cls.get_trade_volume_field_name() in keys:
         
             # Date time
             date_time = float(raw[cls.get_trades_timestamp_field_name()])
@@ -139,8 +144,7 @@ class ExchGwApiQuoine(RESTfulApiSocket):
         :param instmt: Instrument
         :return: Object L2Depth
         """
-        link = cls.get_order_book_link(instmt)
-        res = cls.request(link)
+        res = cls.request(cls.get_order_book_link(instmt))
         if len(res) > 0:
             return cls.parse_l2_depth(instmt=instmt,
                                        raw=res)
@@ -155,48 +159,28 @@ class ExchGwApiQuoine(RESTfulApiSocket):
         :param trade_id: Trade id
         :return: List of trades
         """
-        exch_trade_id = int(instmt.get_exch_trade_id())
-        page_limit = 10 if exch_trade_id > 0 else 1
+        link = cls.get_trades_link(instmt)
+        res = cls.request(link)
         trades = []
-        is_continue = True
-
-        for page in range(1, page_limit+1):
-            link = cls.get_trades_link(instmt, page)
-            res = cls.request(link)
-            if len(res) > 0:
-                if 'models' not in res.keys():
-                    break
-                res = res['models']
-                for t in res:
-                    trade = cls.parse_trade(instmt=instmt,
-                                             raw=t)
-                    if int(trade.trade_id) > exch_trade_id:
-                        trades.append(trade)
-                    else:
-                        is_continue = False
-                        break
-
-                if not is_continue:
-                    break
+        if len(res) > 0:
+            for t in res:
+                trade = cls.parse_trade(instmt=instmt,
+                                         raw=t)
+                trades.append(trade)
 
         return trades
 
 
-class ExchGwQuoine(ExchangeGateway):
-    # static variable to control to request rate
-    num_of_connections = 0
-    num_of_connections_lock = threading.Lock()
-    extra_waiting_sec = 0.5
-    
+class ExchGwBtcc(ExchangeGateway):
     """
-    Exchange gateway
+    Exchange gateway BTCC
     """
     def __init__(self, db_client):
         """
         Constructor
         :param db_client: Database client
         """
-        ExchangeGateway.__init__(self, ExchGwApiQuoine(), db_client)
+        ExchangeGateway.__init__(self, ExchGwBtccRestfulApi(), db_client)
 
     @classmethod
     def get_exchange_name(cls):
@@ -204,17 +188,13 @@ class ExchGwQuoine(ExchangeGateway):
         Get exchange name
         :return: Exchange name string
         """
-        return 'Quoine'
+        return 'BTCC'
 
     def get_order_book_worker(self, instmt):
         """
         Get order book worker
         :param instmt: Instrument
         """
-        ExchGwQuoine.num_of_connections_lock.acquire()
-        ExchGwQuoine.num_of_connections += 1
-        Logger.info(self.__class__.__name__, "Current number of connections = %d" % ExchGwQuoine.num_of_connections)
-        ExchGwQuoine.num_of_connections_lock.release()
         instmt.set_order_book_id(self.get_order_book_init(instmt))
 
         while True:
@@ -227,19 +207,13 @@ class ExchGwQuoine(ExchangeGateway):
                     self.insert_order_book(instmt)
             except Exception as e:
                 Logger.error(self.__class__.__name__, "Error in order book: %s" % e)
-            ExchGwQuoine.num_of_connections_lock.acquire()
-            time.sleep(ExchGwQuoine.num_of_connections + ExchGwQuoine.extra_waiting_sec)
-            ExchGwQuoine.num_of_connections_lock.release()
+            time.sleep(1)
 
     def get_trades_worker(self, instmt):
         """
         Get order book worker thread
         :param instmt: Instrument name
         """
-        ExchGwQuoine.num_of_connections_lock.acquire()
-        ExchGwQuoine.num_of_connections += 1
-        Logger.info(self.__class__.__name__, "Current number of connections = %d" % ExchGwQuoine.num_of_connections)
-        ExchGwQuoine.num_of_connections_lock.release()        
         trade_id, exch_trade_id = self.get_trades_init(instmt)
         instmt.set_trade_id(trade_id)
         instmt.set_exch_trade_id(exch_trade_id)
@@ -267,9 +241,7 @@ class ExchGwQuoine(ExchangeGateway):
             if not instmt.get_recovered():
                 instmt.set_recovered(True)
 
-            ExchGwQuoine.num_of_connections_lock.acquire()
-            time.sleep(ExchGwQuoine.num_of_connections + ExchGwQuoine.extra_waiting_sec)
-            ExchGwQuoine.num_of_connections_lock.release()
+            time.sleep(1)
 
     def start(self, instmt):
         """
@@ -289,16 +261,169 @@ class ExchGwQuoine(ExchangeGateway):
         t2 = threading.Thread(target=partial(self.get_trades_worker, instmt))
         t2.start()
         return [t1, t2]
+
+
+class ExchGwBtccSpotRestfulApi(ExchGwBtccRestfulApi):
+    """
+    Exchange gateway Spot Instrument RESTful API
+    """
+    def __init__(self):
+        ExchGwBtccRestfulApi.__init__(self)
+        
+    @classmethod
+    def get_timestamp_offset(cls):
+        return 1
+        
+    @classmethod
+    def get_order_book_timestamp_field_name(cls):
+        return 'date'
+        
+    @classmethod
+    def get_trades_timestamp_field_name(cls):
+        return 'date'
+    
+    @classmethod
+    def get_bids_field_name(cls):
+        return 'bids'
+        
+    @classmethod
+    def get_asks_field_name(cls):
+        return 'asks'
+        
+    @classmethod
+    def get_trade_side_field_name(cls):
+        return 'type'
+        
+    @classmethod
+    def get_trade_id_field_name(cls):
+        return 'tid'
+        
+    @classmethod
+    def get_trade_price_field_name(cls):
+        return 'price'        
+        
+    @classmethod
+    def get_trade_volume_field_name(cls):
+        return 'amount'        
+        
+    @classmethod
+    def get_order_book_link(cls, instmt):
+        return "https://data.btcchina.com/data/orderbook?limit=5&market=%s" % instmt.get_instmt_code()
+
+    @classmethod
+    def get_trades_link(cls, instmt):
+        if int(instmt.get_exch_trade_id()) > 0:
+            return "https://data.btcchina.com/data/historydata?market=%s&since=%s" % \
+                (instmt.get_instmt_code(), instmt.get_exch_trade_id())
+        else:
+            return "https://data.btcchina.com/data/historydata?limit=100&market=%s" % \
+                (instmt.get_instmt_code())        
+
+class ExchGwBtccFutureRestfulApi(ExchGwBtccRestfulApi):
+    """
+    Exchange gateway Spot Instrument RESTful API
+    """
+    def __init__(self):
+        ExchGwBtccRestfulApi.__init__(self)
+        
+    @classmethod
+    def get_timestamp_offset(cls):
+        return 1000
+        
+    @classmethod
+    def get_order_book_timestamp_field_name(cls):
+        return 'date'
+        
+    @classmethod
+    def get_trades_timestamp_field_name(cls):
+        return 'Timestamp'
+    
+    @classmethod
+    def get_bids_field_name(cls):
+        return 'bids'
+        
+    @classmethod
+    def get_asks_field_name(cls):
+        return 'asks'
+        
+    @classmethod
+    def get_trade_side_field_name(cls):
+        return 'Side'
+        
+    @classmethod
+    def get_trade_id_field_name(cls):
+        return 'Id'
+        
+    @classmethod
+    def get_trade_price_field_name(cls):
+        return 'Price'        
+        
+    @classmethod
+    def get_trade_volume_field_name(cls):
+        return 'Quantity'        
+        
+    @classmethod
+    def get_order_book_link(cls, instmt):
+        return "https://pro-data.btcc.com/data/pro/orderbook?limit=5&symbol=%s" % instmt.get_instmt_code()
+
+    @classmethod
+    def get_trades_link(cls, instmt):
+        if int(instmt.get_exch_trade_id()) > 0:
+            return "https://pro-data.btcc.com/data/pro/historydata?symbol=%s&since=%s" % \
+                (instmt.get_instmt_code(), instmt.get_exch_trade_id())
+        else:
+            return "https://pro-data.btcc.com/data/pro/historydata?limit=100&symbol=%s" % \
+                (instmt.get_instmt_code())        
+
+
+class ExchGwBtccSpot(ExchGwBtcc):
+    """
+    Exchange gateway BTCC-Spot
+    """
+    def __init__(self, db_client):
+        """
+        Constructor
+        :param db_client: Database client
+        """
+        ExchangeGateway.__init__(self, ExchGwBtccSpotRestfulApi(), db_client)
+
+    @classmethod
+    def get_exchange_name(cls):
+        """
+        Get exchange name
+        :return: Exchange name string
+        """
+        return 'BTCC_Spot'
+        
+        
+class ExchGwBtccFuture(ExchGwBtcc):
+    """
+    Exchange gateway BTCC-Future
+    """
+    def __init__(self, db_client):
+        """
+        Constructor
+        :param db_client: Database client
+        """
+        ExchangeGateway.__init__(self, ExchGwBtccFutureRestfulApi(), db_client)
+
+    @classmethod
+    def get_exchange_name(cls):
+        """
+        Get exchange name
+        :return: Exchange name string
+        """
+        return 'BTCC_Future'        
         
         
 if __name__ == '__main__':
     Logger.init_log()
-    exchange_name = 'Quoine'
-    instmt_name = 'BTCUSD'
-    instmt_code = '1'
+    exchange_name = 'BTCC_Spot'
+    instmt_name = 'BTCCNY'
+    instmt_code = 'btccny'
     instmt = Instrument(exchange_name, instmt_name, instmt_code)    
     db_client = SqlClientTemplate()
-    exch = ExchGwQuoine(db_client)
+    exch = ExchGwBtccSpot(db_client)
     instmt.set_l2_depth(L2Depth(5))
     instmt.set_prev_l2_depth(L2Depth(5))
     instmt.set_order_book_table_name(exch.get_order_book_table_name(instmt.get_exchange_name(),
@@ -306,5 +431,4 @@ if __name__ == '__main__':
     instmt.set_trades_table_name(exch.get_trades_table_name(instmt.get_exchange_name(),
                                                             instmt.get_instmt_name()))
     instmt.set_recovered(False)    
-    # exch.get_order_book_worker(instmt)
-    exch.get_trades_worker(instmt)
+    exch.get_order_book_worker(instmt)
