+++ befh/exch_ws_template.py
@@ -1,8 +1,9 @@
-from bitcoinexchangefh.ws_api_socket import WebSocketApiClient
-from bitcoinexchangefh.market_data import L2Depth, Trade
-from bitcoinexchangefh.exchange import ExchangeGateway
-from bitcoinexchangefh.instrument import Instrument
-from bitcoinexchangefh.util import Logger
+from befh.ws_api_socket import WebSocketApiClient
+from befh.market_data import L2Depth, Trade
+from befh.exchange import ExchangeGateway
+from befh.instrument import Instrument
+from befh.sql_client_template import SqlClientTemplate
+from befh.util import Logger
 import time
 import threading
 import json
@@ -10,7 +11,7 @@ from functools import partial
 from datetime import datetime
 
 
-class ExchGwOkCoinWs(WebSocketApiClient):
+class ExchGwApiTemplate(WebSocketApiClient):
     """
     Exchange socket
     """
@@ -18,10 +19,14 @@ class ExchGwOkCoinWs(WebSocketApiClient):
         """
         Constructor
         """
-        WebSocketApiClient.__init__(self, 'ExchGwOkCoin')
+        WebSocketApiClient.__init__(self, 'Template')
         
     @classmethod
     def get_order_book_timestamp_field_name(cls):
+        return 'timestamp'
+        
+    @classmethod
+    def get_trades_timestamp_field_name(cls):
         return 'timestamp'
     
     @classmethod
@@ -33,17 +38,33 @@ class ExchGwOkCoinWs(WebSocketApiClient):
         return 'asks'
         
     @classmethod
+    def get_trade_side_field_name(cls):
+        return 'side'
+        
+    @classmethod
+    def get_trade_id_field_name(cls):
+        return 'trdMatchID'
+        
+    @classmethod
+    def get_trade_price_field_name(cls):
+        return 'price'        
+        
+    @classmethod
+    def get_trade_volume_field_name(cls):
+        return 'size'   
+        
+    @classmethod
     def get_link(cls):
-        return 'wss://real.okcoin.com:10440/websocket/okcoinapi'
-
+        return 'wss://www.bitmex.com/realtime'
+        
     @classmethod
     def get_order_book_subscription_string(cls, instmt):
-        return json.dumps({"event":"addChannel", "channel": instmt.get_order_book_channel_id()})
-
+        return json.dumps({"op":"subscribe", "args": ["orderBook10:%s" % instmt.get_instmt_code()]})
+        
     @classmethod
     def get_trades_subscription_string(cls, instmt):
-        return json.dumps({"event":"addChannel", "channel": instmt.get_trades_channel_id()})
-
+        return json.dumps({"op":"subscribe", "args": ["trade:%s" % instmt.get_instmt_code()]})
+            
     @classmethod
     def parse_l2_depth(cls, instmt, raw):
         """
@@ -58,8 +79,9 @@ class ExchGwOkCoinWs(WebSocketApiClient):
            cls.get_asks_field_name() in keys:
             
             # Date time
-            timestamp = float(raw[cls.get_order_book_timestamp_field_name()])/1000.0
-            l2_depth.date_time = datetime.utcfromtimestamp(timestamp).strftime("%Y%m%d %H:%M:%S.%f")
+            timestamp = raw[cls.get_order_book_timestamp_field_name()]
+            timestamp = timestamp.replace('T', ' ').replace('Z', '').replace('-' , '')
+            l2_depth.date_time = timestamp
             
             # Bids
             bids = raw[cls.get_bids_field_name()]
@@ -77,9 +99,9 @@ class ExchGwOkCoinWs(WebSocketApiClient):
         else:
             raise Exception('Does not contain order book keys in instmt %s-%s.\nOriginal:\n%s' % \
                 (instmt.get_exchange_name(), instmt.get_instmt_name(), \
-                 raw))        
-
-        return l2_depth
+                 raw))
+        
+        return l2_depth        
 
     @classmethod
     def parse_trade(cls, instmt, raw):
@@ -89,21 +111,39 @@ class ExchGwOkCoinWs(WebSocketApiClient):
         :return:
         """
         trade = Trade()
-        trade_id = raw[0]
-        trade_price = float(raw[1])
-        trade_volume = float(raw[2])
-        timestamp = raw[3]
-        trade_side = raw[4]
+        keys = list(raw.keys())
         
-        trade.trade_id = trade_id + timestamp
-        trade.trade_price = trade_price
-        trade.trade_volume = trade_volume
-        trade.trade_side = Trade.parse_side(trade_side)
+        if cls.get_trades_timestamp_field_name() in keys and \
+           cls.get_trade_id_field_name() in keys and \
+           cls.get_trade_side_field_name() in keys and \
+           cls.get_trade_price_field_name() in keys and \
+           cls.get_trade_volume_field_name() in keys:
+        
+            # Date time
+            timestamp = raw[cls.get_trades_timestamp_field_name()]
+            timestamp = timestamp.replace('T', ' ').replace('Z', '').replace('-' , '')
+            trade.date_time = timestamp
+            
+            # Trade side
+            trade.trade_side = Trade.parse_side(raw[cls.get_trade_side_field_name()])
+                
+            # Trade id
+            trade.trade_id = raw[cls.get_trade_id_field_name()]
+            
+            # Trade price
+            trade.trade_price = raw[cls.get_trade_price_field_name()]            
+            
+            # Trade volume
+            trade.trade_volume = raw[cls.get_trade_volume_field_name()]                        
+        else:
+            raise Exception('Does not contain trade keys in instmt %s-%s.\nOriginal:\n%s' % \
+                (instmt.get_exchange_name(), instmt.get_instmt_name(), \
+                 raw))        
 
-        return trade
+        return trade        
 
 
-class ExchGwOkCoin(ExchangeGateway):
+class ExchGwTemplate(ExchangeGateway):
     """
     Exchange gateway
     """
@@ -112,7 +152,7 @@ class ExchGwOkCoin(ExchangeGateway):
         Constructor
         :param db_client: Database client
         """
-        ExchangeGateway.__init__(self, ExchGwOkCoinWs(), db_client)
+        ExchangeGateway.__init__(self, ExchGwApiTemplate(), db_client)
 
     @classmethod
     def get_exchange_name(cls):
@@ -120,7 +160,7 @@ class ExchGwOkCoin(ExchangeGateway):
         Get exchange name
         :return: Exchange name string
         """
-        return 'OkCoin'
+        return 'Template'
 
     def on_open_handler(self, instmt, ws):
         """
@@ -131,22 +171,6 @@ class ExchGwOkCoin(ExchangeGateway):
         Logger.info(self.__class__.__name__, "Instrument %s is subscribed in channel %s" % \
                   (instmt.get_instmt_code(), instmt.get_exchange_name()))
         if not instmt.get_subscribed():
-            instmt_code_split = instmt.get_instmt_code().split('_')
-            if len(instmt_code_split) == 3:
-                # Future instruments
-                instmt.set_order_book_channel_id("ok_sub_%s_%s_depth_%s_20" % \
-                                                 (instmt_code_split[0],
-                                                  instmt_code_split[1],
-                                                  instmt_code_split[2]))
-                instmt.set_trades_channel_id("ok_sub_%s_%s_trade_%s" % \
-                                               (instmt_code_split[0],
-                                                instmt_code_split[1],
-                                                instmt_code_split[2]))
-            else:
-                # Spot instruments
-                instmt.set_order_book_channel_id("ok_sub_%s_depth_20" % instmt.get_instmt_code())
-                instmt.set_trades_channel_id("ok_sub_%s_trades" % instmt.get_instmt_code())
-
             ws.send(self.api_socket.get_order_book_subscription_string(instmt))
             ws.send(self.api_socket.get_trades_subscription_string(instmt))
             instmt.set_subscribed(True)
@@ -157,43 +181,45 @@ class ExchGwOkCoin(ExchangeGateway):
         :param instmt: Instrument
         :param ws: Web socket
         """
-        Logger.info(self.__class__.__name__, "Instrument %s is unsubscribed in channel %s" % \
+        Logger.info(self.__class__.__name__, "Instrument %s is subscribed in channel %s" % \
                   (instmt.get_instmt_code(), instmt.get_exchange_name()))
         instmt.set_subscribed(False)
 
-    def on_message_handler(self, instmt, messages):
+    def on_message_handler(self, instmt, message):
         """
         Incoming message handler
         :param instmt: Instrument
         :param message: Message
         """
-        for message in messages:
-            keys = message.keys()
-            if 'channel' in keys:
-                if 'data' in keys:
-                    if message['channel'] == instmt.get_order_book_channel_id():
-                        data = message['data']
+        keys = message.keys()
+        if 'info' in keys:
+            Logger.info(self.__class__.__name__, message['info'])
+        elif 'subscribe' in keys:
+            Logger.info(self.__class__.__name__, 'Subscription of %s is %s' % \
+                        (message['request']['args'], \
+                         'successful' if message['success'] else 'failed'))
+        elif 'table' in keys:
+            if message['table'] == 'trade':
+                for trade_raw in message['data']:
+                    if trade_raw["symbol"] == instmt.get_instmt_code():
+                        # Filter out the initial subscriptions
+                        trade = self.api_socket.parse_trade(instmt, trade_raw)
+                        if trade.trade_id != instmt.get_exch_trade_id():
+                            instmt.incr_trade_id()
+                            instmt.set_exch_trade_id(trade.trade_id)
+                            self.insert_trade(instmt, trade)
+            elif message['table'] == 'orderBook10':
+                for data in message['data']:
+                    if data["symbol"] == instmt.get_instmt_code():
                         instmt.set_prev_l2_depth(instmt.get_l2_depth().copy())
                         self.api_socket.parse_l2_depth(instmt, data)
-
-                        # Insert only if the first 5 levels are different
                         if instmt.get_l2_depth().is_diff(instmt.get_prev_l2_depth()):
                             instmt.incr_order_book_id()
                             self.insert_order_book(instmt)
-
-                    elif message['channel'] == instmt.get_trades_channel_id():
-                        for trade_raw in message['data']:
-                            trade = self.api_socket.parse_trade(instmt, trade_raw)
-                            if trade.trade_id != instmt.get_exch_trade_id():
-                                instmt.incr_trade_id()
-                                instmt.set_exch_trade_id(trade.trade_id)
-                                self.insert_trade(instmt, trade)
-
-                elif 'success' in keys:
-                    Logger.info(self.__class__.__name__, "Subscription to channel %s is %s" \
-                        % (message['channel'], message['success']))
             else:
-                Logger.info(self.__class__.__name__, ' - ' + json.dumps(message))
+                Logger.info(self.__class__.__name__, json.dumps(message,indent=2))
+        else:
+            Logger.error(self.__class__.__name__, "Unrecognised message:\n" + json.dumps(message))
 
     def start(self, instmt):
         """
@@ -201,8 +227,8 @@ class ExchGwOkCoin(ExchangeGateway):
         :param instmt: Instrument
         :return List of threads
         """
-        instmt.set_prev_l2_depth(L2Depth(20))
-        instmt.set_l2_depth(L2Depth(20))
+        instmt.set_l2_depth(L2Depth(10))
+        instmt.set_prev_l2_depth(L2Depth(10))
         instmt.set_order_book_table_name(self.get_order_book_table_name(instmt.get_exchange_name(),
                                                                        instmt.get_instmt_name()))
         instmt.set_trades_table_name(self.get_trades_table_name(instmt.get_exchange_name(),
@@ -215,4 +241,15 @@ class ExchGwOkCoin(ExchangeGateway):
                                         on_message_handler=partial(self.on_message_handler, instmt),
                                         on_open_handler=partial(self.on_open_handler, instmt),
                                         on_close_handler=partial(self.on_close_handler, instmt))]
+                                        
+
+if __name__ == '__main__':
+    exchange_name = 'Template'
+    instmt_name = 'XBTUSD'
+    instmt_code = 'XBTH17'
+    instmt = Instrument(exchange_name, instmt_name, instmt_code)
+    db_client = SqlClientTemplate()
+    Logger.init_log()
+    exch = ExchGwTemplate(db_client)
+    td = exch.start(instmt)
