+++ befh/test/test_sqlite_client.py
